requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
sentence-transformers
faiss-cpu
numpy
jinja2
